+++ src/controllers/profesores.c.js
@@ -1,5 +1,6 @@
 const materiasModel = require("../models/materias.m.js");
 const profesoresModel = require("../models/profesores.m.js");
+const { autenticacion } = require("./jwt/autenticacion.js");
 
 class profesoresControllers {
   async listar() {
@@ -33,18 +34,22 @@ class profesoresControllers {
   async agregar(profesor) {
     return new Promise(async (resolve, reject) => {
       try {
-        const verificacionExiste = await profesoresModel.find({nombre: profesor.nombre, cedula: profesor.cedula}); // Validamos que no se repitan los profesores
+        const acceso = await autenticacion(profesor.token, ['director'])
+        if (acceso != 'acceso permitido') {
+          return reject(acceso)
+        }
+        const verificacionExiste = await profesoresModel.find({ nombre: profesor.nombre, cedula: profesor.cedula }); // Validamos que no se repitan los profesores
         if (verificacionExiste.length > 0) {
-            return reject("Ya esta registrado el profesor");
+          return reject("Ya esta registrado el profesor");
         }
         const data = {
-            nombre: profesor.nombre,
-            telefono: Number(profesor.telefono),
-            cedula: Number(profesor.cedula),
+          nombre: profesor.nombre,
+          telefono: Number(profesor.telefono),
+          cedula: Number(profesor.cedula),
         } // Creamos el documento con los tipos de datos correctos
         const datos = await profesoresModel.create(data);
         if (datos) {
-            return resolve(datos)
+          return resolve(datos)
         }
         return reject("No se pudo agregar el profesor")
       } catch (error) {
@@ -56,28 +61,32 @@ class profesoresControllers {
   async actualizar(id, profesor) {
     return new Promise(async (resolve, reject) => {
       try {
+        const acceso = await autenticacion(profesor.token, ['director'])
+        if (acceso != 'acceso permitido') {
+          return reject(acceso)
+        }
         const verificacionExisteId = await profesoresModel.findById(id); // Validamos que exista el profesor
         if (!verificacionExisteId) {
-            return reject("No existe el profesor")
+          return reject("No existe el profesor")
         }
-        const verificacionExiste = await profesoresModel.find({nombre: profesor.nombre, cedula: Number(profesor.cedula)}); // Validamos que no se repitan
+        const verificacionExiste = await profesoresModel.find({ nombre: profesor.nombre, cedula: Number(profesor.cedula) }); // Validamos que no se repitan
         console.log(verificacionExiste)
         if (verificacionExiste.length > 0) {
-            return reject("Ya existe este profesor");
+          return reject("Ya existe este profesor");
         }
         const data = {
-            nombre: profesor.nombre,
-            cedula: Number(profesor.cedula),
-            telefono: Number(profesor.telefono)
+          nombre: profesor.nombre,
+          cedula: Number(profesor.cedula),
+          telefono: Number(profesor.telefono)
         } // Creamos el documento con los tipos de datos correctos
         const datos = await profesoresModel.findByIdAndUpdate(id, data);
         if (datos) {
-            return resolve({
-                _id: datos._id,
-                nombre: profesor.nombre,
-                cedula: Number(profesor.cedula),
-                telefono: Number(profesor.telefono)
-            })
+          return resolve({
+            _id: datos._id,
+            nombre: profesor.nombre,
+            cedula: Number(profesor.cedula),
+            telefono: Number(profesor.telefono)
+          })
         }
         return reject("No se pudo editar el profesor")
       } catch (error) {
@@ -86,17 +95,21 @@ class profesoresControllers {
     });
   }
 
-  async eliminar(id) {
+  async eliminar(id, profesor) {
     return new Promise(async (resolve, reject) => {
       try {
+        const acceso = await autenticacion(profesor.token, ['director'])
+        if (acceso != 'acceso permitido') {
+          return reject(acceso)
+        }
         const verificacionExisteId = await profesoresModel.findById(id); // Validamos que exista el profesor
         if (!verificacionExisteId) {
-            return reject("No existe el profesor")
+          return reject("No existe el profesor")
         }
-        await materiasModel.updateMany({profesorId: id}, {profesorId: "Sin profesor"})
+        await materiasModel.updateMany({ profesorId: id }, { profesorId: "Sin profesor" })
         const datos = await profesoresModel.findByIdAndDelete(id); // Eliminamos la materia
         if (datos) {
-            return resolve(datos)
+          return resolve(datos)
         }
         return reject("No se pudo eliminar la materia")
       } catch (error) {
